問い合わせformとdialogを分離させた
